% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceden_query.R
\name{ceden_query}
\alias{ceden_query}
\title{Download CEDEN data via an API}
\usage{
ceden_query(service, query_parameters,
  base_URI = "https://testcedenwebservices.waterboards.ca.gov:9267",
  userName = "", password = "")
}
\arguments{
\item{service}{The CEDEN web service}
}
\value{
dataframe with the data specified in the service and query_parameters
}
\description{
This function provides an interface with CEDEN webservices to download data
}
\examples{
ceden_query(service = 'cedenwaterqualitymonitoringstationslist', query_parameters = '"filter":[{"sampleDateMin":"1/1/2015","sampleDateMax":"4/1/2015"}],"top":1000')
}
\keyword{API}
\keyword{CEDEN}
\keyword{California}
